+++ view/frontend/requirejs-config.js
@@ -1,7 +1,8 @@
 var config = {
     map: {
         '*': {
-            KlaviyoCustomerData: 'Klaviyo_Reclaim/js/customer'
+            KlaviyoCustomerData: 'Klaviyo_Reclaim/js/customer',
+            'Magento_Checkout/js/model/shipping-save-processor/default': 'Klaviyo_Reclaim/js/model/shipping-save-processor/default'
         }
     }
 };